Implemented DummyInferenceEngine with CLI support and fixed latency test issues.
- Added CLI argument support `--inference-engine dummy` to use DummyInferenceEngine.
- Fixed intermittent failure in latency test by allowing a small timing tolerance.
- Refactored DummyInferenceEngine to ensure correct output and behavior.
- Refactored leftover code in test_dummy_inference_engine and DummyInferenceEngine
