+++ exo/inference/DummyInferenceEngine.py
@@ -3,7 +3,7 @@ import asyncio
 import numpy as np  # Import numpy for generating random outputs
 
 class DummyInferenceEngine:
-    def __init__(self, output_type="static", output_value=None, output_shape=(1,), latency_mean=0.1, latency_stddev=0.01):
+    def __init__(self, output_type="static", output_value=None, output_shape=(1,), latency_mean=0.1, latency_stddev=0.1):
         self.output_type = output_type
         self.output_value = output_value
         self.output_shape = output_shape
@@ -30,20 +30,3 @@ class DummyInferenceEngine:
         # Simulate and return the latency
         return max(0, random.normalvariate(self.latency_mean, self.latency_stddev))
 
-
-# Example usage and testing
-async def test_dummy_engine():
-    dummy_engine = DummyInferenceEngine(output_type="random", output_shape=(2, 2), latency_mean=0.5, latency_stddev=0.1)
-    
-    # Simulate inference
-    output = await dummy_engine.run_inference()
-    latency = await dummy_engine.get_latency()
-    
-    assert isinstance(output, list), "Output should be a list."
-    assert isinstance(latency, float), "Latency should be a float."
-    
-    print(f"Simulated Inference Output: {output}")
-    print(f"Simulated Inference Latency: {latency}s")
-
-# Run the test
-# asyncio.run(test_dummy_engine())
